Fix clippy, benches
