+++ Cargo.toml
@@ -1,4 +1,5 @@
 [workspace]
+resolver = "2"
 members = [
     "shogi_legality_lite",
     "shogi_legality_lite_c",
