+++ shogi_legality_lite/Cargo.toml
@@ -3,7 +3,7 @@ name = "shogi_legality_lite"
 version = "0.1.2"
 authors = ["Rust shogi crates developers"]
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.74"
 description = "Legality checking in shogi without tables"
 documentation = "https://docs.rs/shogi_legality_lite"
 readme = "README.md"
@@ -25,9 +25,19 @@ std = ["alloc", "shogi_core/std"]
 crate-type = [
     "rlib",
 ]
+bench = false
 
 [dependencies]
 shogi_core = { version = "0.1.3", default-features = false }
 
 [dev-dependencies]
 shogi_usi_parser = { version = "=0.1.0", default-features = false }
+criterion = "0.5"
+
+[[bench]]
+name = "perft"
+harness = false
+
+[[bench]]
+name = "mate_solver"
+harness = false
