+++ shogi_legality_lite/src/lib.rs
@@ -1,10 +1,6 @@
 #![cfg_attr(not(test), no_std)] // Forbids using std::*.
 #![cfg_attr(docsrs, feature(doc_cfg))]
-#![cfg_attr(bench, feature(test))]
 #![doc = include_str!("../README.md")]
-
-#[cfg(bench)]
-extern crate test;
 
 #[cfg(feature = "alloc")]
 extern crate alloc;
